requirements.txt
asgiref==3.5.2
click==8.1.3
colorama==0.4.5
Django==4.0.5
django-model-utils==4.2.0
django-notifications-hq==1.7.0
djangorestframework==3.12.4
joblib==1.1.0
jsonfield==3.1.0
nltk==3.7
numpy==1.22.4
psycopg2==2.9.3
pytz==2022.1
regex==2022.6.2
scikit-learn==1.1.1
scipy==1.8.1
sklearn==0.0
sqlparse==0.4.2
swapper==1.3.0
threadpoolctl==3.1.0
tqdm==4.64.0
tzdata==2022.1
